grpc-fd: rename grpc-python-module
The name did not say much and was also misleading.  We chose a name
that hopefully makes it clear what it actually does when it shows up
in a process explorer.

(cherry picked from commit fbfa6846a425fd15e2bbb3a3e6530d81a5de3aa3)
