+++ debian/bareos-filedaemon-grpc-python3-plugin.install.in
@@ -1,3 +1,3 @@
 @plugindir@/grpc-fd.so
-@plugindir@/grpc/grpc-python-module
+@plugindir@/grpc/bareos-grpc-fd-plugin-bridge
 @plugindir@/grpc/grpc-test-module
