+++ docs/manuals/source/TasksAndConcepts/Plugins/FileDaemonPlugins/GrpcPlugin.rst.inc
@@ -18,10 +18,10 @@ This has multiple upsides for users, such as:
 This plugin on its own is not very useful.  Its only a bridge between the bareos core and
 the actual plugin doing the work.
 
-The plugin comes with two executables, `grpc-test-module` and `grpc-python-module`, which allow you
+The plugin comes with two executables, `grpc-test-module` and `bareos-grpc-fd-plugin-bridge`, which allow you
 to make use of this bridge.
 As the name suggests, `grpc-test-module` is a simple module that can be used to test that the bridge is working.
-`grpc-python-module` on the other hand is a handy little executable that can be used to load and run
+`bareos-grpc-fd-plugin-bridge` on the other hand is a handy little executable that can be used to load and run
 normal bareos plugins (including python plugins) in a separate process.
 
 Status of the Grpc Plugin
@@ -31,7 +31,7 @@ This plugin is still in an **experimental** phase.  The API between core and plu
 
 Installation of the Grpc Plugin
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-The grpc plugin, together with the `grpc-test-module` and `grpc-python-module`, can be installed with the
+The grpc plugin, together with the `grpc-test-module` and `bareos-grpc-fd-plugin-bridge`, can be installed with the
 **bareos-filedaemon-gprc-plugin** package.
 
 Configuration of the Grpc Plugin
@@ -61,10 +61,10 @@ different jobs.
      [core] <-> [grpc-fd] : Bareos Plugin API
    }
    package "Inferior Process" {
-     [grpc-python-module] <-> [python-fd] : Bareos Plugin API
+     [bareos-grpc-fd-plugin-bridge] <-> [python-fd] : Bareos Plugin API
      [python-fd] <-> [python files] : Python Plugin API
    }
-   [grpc-fd] <--> [grpc-python-module] : Grpc Plugin API
+   [grpc-fd] <--> [bareos-grpc-fd-plugin-bridge] : Grpc Plugin API
    @enduml
 
 This plugin does not take any options, but expects its arguments to be a
@@ -84,7 +84,7 @@ Example
          Signature = XXH128
        }
        Plugin = "grpc"
-                ":grpc-python-module"
+                ":bareos-grpc-fd-plugin-bridge"
                 ":python3"
                 ":module_name=pyplug"
                 ":arg1=val1"
