+++ docs/manuals/source/include/autogenerated/bareos-fd-config-schema.json
@@ -374,7 +374,7 @@
         "GrpcModule": {
           "datatype": "STRING",
           "code": 0,
-          "default_value": "grpc-compat-module",
+          "default_value": "bareos-grpc-fd-plugin-bridge",
           "equals": true,
           "versions": "25.0.0-",
           "description": "The grpc module to use for grpc fallback."
@@ -681,7 +681,7 @@
         "GrpcModule": {
           "datatype": "STRING",
           "code": 0,
-          "default_value": "grpc-compat-module",
+          "default_value": "bareos-grpc-fd-plugin-bridge",
           "equals": true,
           "versions": "25.0.0-",
           "description": "The grpc module to use for grpc fallback."
