+++ systemtests/tests/auto-grpc/etc/bareos/bareos-fd.d/client/myself.conf.in
@@ -4,7 +4,7 @@ Client {
   Working Directory =  "@working_dir@"
   FD Port = @fd_port@
 
-  Grpc Module = grpc-python-module
+  Grpc Module = bareos-grpc-fd-plugin-bridge
 
   Plugin Directory = "@FD_PLUGINS_DIR_TO_TEST@"
   Plugin Names = "grpc"
