+++ systemtests/tests/grpc-fd/etc/bareos/bareos-dir.d/fileset/SelfTest.conf.in
@@ -32,7 +32,7 @@ FileSet {
       Acl Support = No
       Xattr Support = No
     }
-    Plugin = "grpc:grpc-python-module:@python_module_name@"
+    Plugin = "grpc:bareos-grpc-fd-plugin-bridge:@python_module_name@"
              ":module_path=@python_plugin_module_src_test_dir@"
              ":module_name=bareos-fd-local-fileset-acl-xattr"
              ":filename=@tmpdir@/file-list"
@@ -48,7 +48,7 @@ FileSet {
       Acl Support = No
       Xattr Support = No
     }
-    Plugin = "grpc:grpc-python-module:@python_module_name@"
+    Plugin = "grpc:bareos-grpc-fd-plugin-bridge:@python_module_name@"
              ":module_path=@python_plugin_module_src_test_dir@"
              ":module_name=bareos-fd-local-fileset"
              ":filename=@tmpdir@/file-list"
